Create problem_9.c
Problem-9 : Building calculator using switch case. 
Algorithm :
1. Start.
2. Read integer type variables x,y and charecter type variable op.
3. Read the values from user and print  the values.
4. Use a switch(op) statement to check which operator was entered.
If op is '+', calculate x+y and print the result. 
If op is '-', calculate x-y and print the result. 
If op is '*', calculate x*y and print the result. 
If op is '/', calculate x/y and print the result. 
Otherwise, print " Incorrect operator".
5. End.